Parametrizacion de todos los alias, reestructuracion leve de la arquitectura y programacion del boton refrescar
