+++ codigo/componentes/navegador/WindowsFormsApp1/CapaModeloNavegador/ConexionMYSQL.cs
@@ -9,7 +9,7 @@ namespace CapaModeloNavegador
 {
     public class ConexionMYSQL
     {
-        private readonly string ConexionODBC = "Dsn=pruebaCapacitacion"; // DSN de odbc
+        private readonly string ConexionODBC = "Dsn=Prueba1"; // DSN de odbc
 
         //retorna conexion cerrada para que el DAO la abra y cierre cuando sea necesario
         public OdbcConnection conexion()
