+++ codigo/componentes/navegador/WindowsFormsApp1/CapaVistaNavegador/Properties/Resources.resx
