+++ codigo/componentes/navegador/WindowsFormsApp1/PruebaEjecucionNavegador/PruebaEjecucionNavegador.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.36324.18
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PruebaEjecucionNavegador", "PruebaEjecucionNavegador\PruebaEjecucionNavegador.csproj", "{3E2584FC-3AA2-4967-A30D-5CEFE0816196}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3E2584FC-3AA2-4967-A30D-5CEFE0816196}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3E2584FC-3AA2-4967-A30D-5CEFE0816196}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3E2584FC-3AA2-4967-A30D-5CEFE0816196}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3E2584FC-3AA2-4967-A30D-5CEFE0816196}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {5DC93FF6-3A82-4B25-9AB9-CE070378F0A1}
+	EndGlobalSection
+EndGlobal
