+++ codigo/componentes/navegador/WindowsFormsApp1/PruebaEjecucionNavegador/PruebaEjecucionNavegador/Program.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace WindowsFormsApp1
+namespace PruebaEjecucionNavegador
 {
-    internal static class Program
+    static class Program
     {
         /// <summary>
         /// Punto de entrada principal para la aplicación.
