+++ codigo/componentes/navegador/WindowsFormsApp1/WindowsFormsApp1.sln
@@ -1,17 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.14.36414.22
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.36324.18
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WindowsFormsApp1", "WindowsFormsApp1\WindowsFormsApp1.csproj", "{C6628A05-1BEC-4276-9F58-210E2A33C380}"
-	ProjectSection(ProjectDependencies) = postProject
-		{44573EC3-7012-4091-81F0-4A5B47C32285} = {44573EC3-7012-4091-81F0-4A5B47C32285}
-		{9CCB68D5-F388-46E4-9B59-97734D28BDC2} = {9CCB68D5-F388-46E4-9B59-97734D28BDC2}
-	EndProjectSection
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CapaControladorNavegador", "CapaControladorNavegador\CapaControladorNavegador.csproj", "{9CCB68D5-F388-46E4-9B59-97734D28BDC2}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CapaModeloNavegador", "CapaModeloNavegador\CapaModeloNavegador.csproj", "{44573EC3-7012-4091-81F0-4A5B47C32285}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CapaVistaNavegador", "CapaVistaNavegador\CapaVistaNavegador.csproj", "{E6A13A61-72FC-4A67-B1AC-6F08C0FD4981}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -19,10 +15,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C6628A05-1BEC-4276-9F58-210E2A33C380}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C6628A05-1BEC-4276-9F58-210E2A33C380}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C6628A05-1BEC-4276-9F58-210E2A33C380}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C6628A05-1BEC-4276-9F58-210E2A33C380}.Release|Any CPU.Build.0 = Release|Any CPU
 		{9CCB68D5-F388-46E4-9B59-97734D28BDC2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9CCB68D5-F388-46E4-9B59-97734D28BDC2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9CCB68D5-F388-46E4-9B59-97734D28BDC2}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -31,6 +23,10 @@ Global
 		{44573EC3-7012-4091-81F0-4A5B47C32285}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{44573EC3-7012-4091-81F0-4A5B47C32285}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{44573EC3-7012-4091-81F0-4A5B47C32285}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E6A13A61-72FC-4A67-B1AC-6F08C0FD4981}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E6A13A61-72FC-4A67-B1AC-6F08C0FD4981}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E6A13A61-72FC-4A67-B1AC-6F08C0FD4981}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E6A13A61-72FC-4A67-B1AC-6F08C0FD4981}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
